Updating tests
Moving testGlossaryLoadFromTerm to single_candidate_page_tests because
we may not always have a glossary term on the home page
